update: Agrega soporte para subir archivos
Se agrega recursos para subir archivos tanto single como multiple aparte
se agregan recursos para obtener información de las facturas de un
cliente determinado

Pendiente: Guardar la información de los archivos en base de datos

+++ model/src/main/java/co/com/business/repository/daos/FacturaDao.java
@@ -1,6 +1,5 @@
 package co.com.business.repository.daos;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import co.com.business.viewobjects.ClienteFacturaVO;
 
@@ -9,7 +8,7 @@ public interface FacturaDao {
 	public ClienteFacturaVO getFacturasCliente(BigInteger cedula);
 	public ClienteFacturaVO getDetallesFactura(BigInteger cedula);
 	public ClienteFacturaVO getDetallesFactura2(BigInteger cedula);
-	public ClienteFacturaVO getFacturaCliente(BigInteger cedula, BigDecimal numeroFactura);	
-	public ClienteFacturaVO getDetalleFactura(BigInteger cedula, BigDecimal numeroFactura);
+	public ClienteFacturaVO getFacturaCliente(BigInteger cedula, Long numeroFactura);	
+	public ClienteFacturaVO getDetalleFactura(BigInteger cedula, Long numeroFactura);
 	
 }
